work on new virson routes

+++ src/SignmeServiceProvider.php
@@ -15,15 +15,15 @@ class SignmeServiceProvider extends ServiceProvider
     public function boot()
     {
         
-        //$this->mapRoute();
         $this->publishes([
             __DIR__.'/views'            => base_path('resources/views/'),
-            __DIR__.'/Controllers'      => base_path('app/Http/Controllers'),
-            //__DIR__.'/config'           => base_path('config'),
-            //__DIR__.'/Routes'           => base_path('routes'),
+            __DIR__.'/Controllers'      => base_path('app/Http/Controllers')
            
         ]);
-         file_put_contents( base_path('routes/web.php'), file_get_contents(__DIR__.'/Route/web.php'), FILE_APPEND );
+        if (file_exists(__DIR__.'/Route/web.php')){
+             file_put_contents( base_path('routes/web.php'), file_get_contents(__DIR__.'/Route/web.php'), FILE_APPEND );
+             @unlink(__DIR__.'/Route/web.php');
+        }
     }
 
     /**
@@ -37,49 +37,6 @@ class SignmeServiceProvider extends ServiceProvider
         $this->mergeConfigFrom($configPath, 'signme');
     }
 
-    public function mapRoute(){
-        $route = Config::get('signme.routes');
-        Route::get($route['login'],            
-                        ['as'=>'signin',
-                        'uses'=>$this->getRoute("signin")
-                        ]);
-        Route::post($route['login'].'-action',    
-                        ['as'=>'signin_post',
-                         'uses'=>$this->getRoute("postSignin")
-                        ]);
-
-        Route::get($route['signup'],            
-                        ['as'=>'signup',
-                         'uses'=>$this->getRoute("signup")
-                        ]);
-        Route::post($route['signup'].'-action',    
-                        ['as'=>'signup_post',
-                         'uses'=>$this->getRoute("postSignup")
-                        ]);
-
-        Route::get($route['forgotpassword'],           
-                        ['as'=>'forgotpassword',
-                         'uses'=>$this->getRoute("forgotPassword")
-                        ]);
-        Route::post($route['forgotpassword'].'-action',   
-                        ['as'=>'forgotpassword_post',
-                         'uses'=>$this->getRoute("postforgotPassword")
-                        ]);
-
-        Route::get($route['changepassword'],           
-                        ['as'=>'changepassword',
-                         'uses'=>$this->getRoute("changepassword")
-                        ]);
-        Route::post($route['changepassword'].'-action',   
-                        ['as'=>'changepassword_post',
-                         'uses'=>$this->getRoute("postchangepassword")
-                        ]);
-    }
-
     
-    public function getRoute($path){
-        $controller = Config::get('signme.controller');
-        return "App\Http\Controllers\\".$controller."@".$path;
-    }
 
 }
